Update jquery.slimmenu.js
Added resize fix for older iOS devices (ios5).  The resize call would fire when the menu would expand, causing it to hide itself.  Fix is to store the width the last time the resize function ran, and check to see if the width has changed; if it has, run the resize; otherwise don't.